+++ src/gui/annotations/qgsmaptoolmodifyannotation.h
@@ -20,7 +20,7 @@
 #include "qgis_gui.h"
 #include "qgis_sip.h"
 #include "qgsannotationitemnode.h"
-#include "qgsmaptooladvanceddigitizing.h"
+#include "qgsannotationmaptool.h"
 #include "qgspointxy.h"
 #include "qgsrectangle.h"
 #include "qobjectuniqueptr.h"
@@ -40,7 +40,7 @@ class QgsSnapIndicator;
  * \note Not available in Python bindings
  * \since QGIS 3.22
  */
-class GUI_EXPORT QgsMapToolModifyAnnotation : public QgsMapToolAdvancedDigitizing
+class GUI_EXPORT QgsMapToolModifyAnnotation : public QgsAnnotationMapTool
 {
     Q_OBJECT
 
@@ -49,7 +49,6 @@ class GUI_EXPORT QgsMapToolModifyAnnotation : public QgsMapToolAdvancedDigitizin
      * Constructor for QgsMapToolModifyAnnotation
      */
     explicit QgsMapToolModifyAnnotation( QgsMapCanvas *canvas, QgsAdvancedDigitizingDockWidget *cadDockWidget );
-
     ~QgsMapToolModifyAnnotation() override;
 
     void deactivate() override;
@@ -86,9 +85,6 @@ class GUI_EXPORT QgsMapToolModifyAnnotation : public QgsMapToolAdvancedDigitizin
     void createHoverBand();
     void createHoveredNodeBand();
     void createSelectedItemBand();
-    const QgsRenderedAnnotationItemDetails *findClosestItemToPoint( const QgsPointXY &mapPoint, const QList<const QgsRenderedAnnotationItemDetails *> &items, QgsRectangle &bounds );
-    QgsAnnotationLayer *annotationLayerFromId( const QString &layerId );
-    QgsAnnotationItem *annotationItemFromId( const QString &layerId, const QString &itemId );
 
     void setHoveredItemFromPoint( const QgsPointXY &mapPoint );
     void setHoveredItem( const QgsRenderedAnnotationItemDetails *item, const QgsRectangle &itemMapBounds );
