+++ src/gui/annotations/qgsmaptoolselectannotationmousehandles.h
@@ -20,14 +20,14 @@
 // We don't want to expose this in the public API
 #define SIP_NO_FILE
 
-#include "qgsmaptoolselectannotation.h"
-#include "qgsgraphicsviewmousehandles.h"
-#include "qgsmapcanvas.h"
-
-#include <QPointer>
 #include <memory>
 
 #include "qgis_gui.h"
+#include "qgsgraphicsviewmousehandles.h"
+#include "qgsmapcanvas.h"
+#include "qgsmaptoolselectannotation.h"
+
+#include <QPointer>
 
 class QGraphicsView;
 class QInputEvent;
