+++ src/gui/maptools/qgsmaptooladvanceddigitizing.h
@@ -20,10 +20,10 @@
 #include <memory>
 
 #include "qgis_gui.h"
+#include "qgsmapmouseevent.h"
 #include "qgsmaptooledit.h"
 #include "qgsreferencedgeometry.h"
 
-class QgsMapMouseEvent;
 class QgsAdvancedDigitizingDockWidget;
 class QgsSnapToGridCanvasItem;
 class QgsSnapIndicator;
